requirements.txt
# Minimal baseline dependencies (pin later if needed)
torch
torchvision
transformers
datasets
pillow
tqdm
pyyaml